# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ServiceOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/setup:
    post:
      operationId: setupUser
      tags: [auth]
      summary: Complete user/company onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupUserRequest"
      responses:
        "200":
          description: Setup complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # Users
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{customerId}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # Jobs
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, cancelled]
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue, cancelled]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{invoiceId}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # Reviews
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get analytics overview
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [analytics]
      summary: Get revenue analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
      responses:
        "200":
          description: Revenue analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueAnalytics"
  /analytics/jobs:
    get:
      operationId: getJobAnalytics
      tags: [analytics]
      summary: Get job analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
      responses:
        "200":
          description: Job analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAnalytics"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [owner, admin, manager, operator]
        phone:
          type: string
          nullable: true
        companyId:
          type: integer
          nullable: true
        company:
          $ref: "#/components/schemas/Company"
          nullable: true
        isOnboarded:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - role
        - isOnboarded
        - createdAt
    SetupUserRequest:
      type: object
      properties:
        companyName:
          type: string
        businessType:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        tier:
          type: string
          enum: [free, independent, pro, franchise]
      required:
        - companyName
        - businessType
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        businessType:
          type: string
        tier:
          type: string
          enum: [free, independent, pro, franchise]
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        stripeCustomerId:
          type: string
          nullable: true
        stripeSubscriptionId:
          type: string
          nullable: true
        maxUsers:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - businessType
        - tier
        - maxUsers
        - isActive
        - createdAt
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        totalJobsCount:
          type: integer
        totalRevenue:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - firstName
        - lastName
        - totalJobsCount
        - totalRevenue
        - createdAt
        - updatedAt
    CreateCustomerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        notes:
          type: string
      required:
        - firstName
        - lastName
    UpdateCustomerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        notes:
          type: string
    CustomerListResponse:
      type: object
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - customers
        - total
        - page
        - limit
    Job:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        customerId:
          type: integer
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        serviceType:
          type: string
          nullable: true
        scheduledStart:
          type: string
          format: date-time
          nullable: true
        scheduledEnd:
          type: string
          format: date-time
          nullable: true
        actualStart:
          type: string
          format: date-time
          nullable: true
        actualEnd:
          type: string
          format: date-time
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        assignedUser:
          $ref: "#/components/schemas/User"
          nullable: true
        estimatedRevenue:
          type: number
          nullable: true
        actualRevenue:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        checkInLat:
          type: number
          nullable: true
        checkInLng:
          type: number
          nullable: true
        checkInTime:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - customerId
        - title
        - status
        - priority
        - createdAt
        - updatedAt
    CreateJobRequest:
      type: object
      properties:
        customerId:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        serviceType:
          type: string
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        assignedToId:
          type: integer
        estimatedRevenue:
          type: number
        notes:
          type: string
      required:
        - customerId
        - title
...
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - jobs
        - total
        - page
        - limit
    Invoice:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        customerId:
          type: integer
        jobId:
          type: integer
          nullable: true
        invoiceNumber:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        subtotal:
          type: number
        taxRate:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        dueDate:
          type: string
          format: date
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - customerId
        - invoiceNumber
        - status
        - subtotal
        - taxRate
        - taxAmount
        - total
        - createdAt
        - updatedAt
    CreateInvoiceRequest:
      type: object
      properties:
        customerId:
          type: integer
        jobId:
          type: integer
        subtotal:
          type: number
        taxRate:
          type: number
        dueDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - customerId
        - subtotal
    UpdateInvoiceRequest:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        subtotal:
          type: number
        taxRate:
          type: number
        dueDate:
          type: string
          format: date
        notes:
          type: string
        paidAt:
          type: string
          format: date-time
    InvoiceListResponse:
      type: object
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - invoices
        - total
        - page
        - limit
...
    AnalyticsOverview:
      type: object
      properties:
        totalJobs:
          type: integer
        completedJobs:
          type: integer
        totalRevenue:
          type: number
        pendingRevenue:
          type: number
        totalCustomers:
          type: integer
        newLeads:
          type: integer
        avgJobValue:
          type: number
        avgRating:
          type: number
        jobCompletionRate:
          type: number
        revenueGrowth:
          type: number
      required:
        - totalJobs
        - completedJobs
        - totalRevenue
        - pendingRevenue
        - totalCustomers
        - newLeads
        - avgJobValue
        - avgRating
        - jobCompletionRate
        - revenueGrowth
    RevenueAnalytics:
      type: object
      properties:
        period:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              revenue:
                type: number
              invoiceCount:
                type: integer
            required:
              - date
              - revenue
              - invoiceCount
        total:
          type: number
        growth:
          type: number
      required:
        - period
        - data
        - total
        - growth
    JobAnalytics:
      type: object
      properties:
        period:
          type: string
        byStatus:
          type: object
          properties:
            scheduled:
              type: integer
            in_progress:
              type: integer
            completed:
              type: integer
            cancelled:
              type: integer
          required:
            - scheduled
            - in_progress
            - completed
            - cancelled
        byServiceType:
          type: array
          items:
            type: object
            properties:
              serviceType:
                type: string
              count:
                type: integer
            required:
              - serviceType
              - count
        avgCompletionTime:
          type: number
      required:
        - period
        - byStatus
        - byServiceType
        - avgCompletionTime